+++ src/main/java/net/mtrop/doom/tools/doommake/ProjectTemplate.java
@@ -1,196 +1,86 @@
 package net.mtrop.doom.tools.doommake;
 
 import java.io.File;
-import java.io.FileOutputStream;
 import java.io.IOException;
-import java.io.InputStream;
+import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.List;
-
-import net.mtrop.doom.struct.io.IOUtils;
-import net.mtrop.doom.tools.common.Common;
+import java.util.Objects;
+import java.util.Set;
 
 /**
- * A project template definition for DoomMake. 
+ * A template descriptor.
  * @author Matthew Tropiano
  */
-public final class ProjectTemplate 
+public class ProjectTemplate implements Comparable<ProjectTemplate>
 {
-	private static final String[] BLANK_RESOURCES = new String[0];
-	
-	/** Template description. */
+	/** Descriptor name. */
+	private String name;
+	/** Descriptor description. */
 	private String description;
-	/** Template file entries. */
-	private List<Entry> entries;
-	
-	/**
-	 * A template entry.
-	 */
-	public static class Entry
-	{
-		private boolean appending;
-		private String outputPath;
-		private String[] resourcePaths;
-		
-		private Entry(boolean appending, String outputPath, String[] resourcePaths) 
-		{
-			super();
-			this.appending = appending;
-			this.outputPath = outputPath;
-			this.resourcePaths = resourcePaths;
-		}
-		
-		private boolean isDirectory()
-		{
-			return resourcePaths == null;
-		}
-	}
-	
-	/**
-	 * Creates an entry for a template directory.
-	 * @param directoryPath the output directory.
-	 * @return the new entry.
-	 */
-	public static Entry dir(String directoryPath)
-	{
-		return new Entry(false, directoryPath, null);
-	}
-	
-	/**
-	 * Creates an entry for a blank template file.
-	 * @param filePath the output file.
-	 * @return the new entry.
-	 */
-	public static Entry file(String filePath)
-	{
-		return new Entry(false, filePath, BLANK_RESOURCES);
-	}
+	/** Order added. */
+	private List<ProjectModule> moduleList;
 
-	/**
-	 * Creates an entry for a template file.
-	 * @param filePath the output file.
-	 * @param resources the series of resources to read and combine into a file.
-	 * @return the new entry.
-	 */
-	public static Entry file(String filePath, String... resources)
-	{
-		return new Entry(false, filePath, resources);
-	}
-	
-	/**
-	 * Creates an entry for a template file that appends to existing files.
-	 * @param filePath the output file.
-	 * @param resources the series of resources to read and combine into a file.
-	 * @return the new entry.
-	 */
-	public static Entry fileAppend(String filePath, String... resources)
-	{
-		return new Entry(true, filePath, resources);
-	}
-	
-	/**
-	 * Creates a new template.
-	 * @param description the template description.
-	 * @param entries the template entries.
-	 * @return the new template.
-	 */
-	public static ProjectTemplate create(String description, Entry... entries)
-	{
-		return new ProjectTemplate(description, entries);
-	}
+	/** Quick lookup for added modules so that included templates are not re-added. */
+	private Set<ProjectModule> moduleSet;
 
-	/**
-	 * Creates a new template.
-	 * @param entries the template entries.
-	 * @return the new template.
-	 */
-	public static ProjectTemplate create(Entry... entries)
+	private ProjectTemplate(String name, String description)
 	{
-		return create(null, entries);
-	}
-
-	// New project template.
-	private ProjectTemplate(String description, Entry... entries)
-	{
-		this.description = description;
-		this.entries = new LinkedList<>();
-		addEntries(entries);
-	}
-	
-	// Add template entries to this one.
-	private void addEntries(Entry... entries)
-	{
-		for (Entry entry : entries)
-			this.entries.add(entry);
-	}
-	
-	// Add template entries to this one.
-	private void addEntries(Iterable<Entry> entries)
-	{
-		for (Entry entry : entries)
-			this.entries.add(entry);
+		this.name = Objects.requireNonNull(name);
+		this.description = Objects.requireNonNull(description);
+		this.moduleList = new LinkedList<>();
+		this.moduleSet = new HashSet<>();
 	}
 	
 	/**
-	 * Adds a template's entries to this one.
-	 * @param template the template to combine with this one.
-	 * @return this template.
+	 * Creates a new {@link ProjectTemplate}.
+	 * @param name the name.
+	 * @param description the description.
+	 * @return a new descriptor.
 	 */
-	public ProjectTemplate add(ProjectTemplate template)
+	public static ProjectTemplate build(String name, String description)
 	{
-		addEntries(template.entries);
-		return this;
+		return new ProjectTemplate(name, description);
 	}
 	
-	/**
-	 * @return this template's description.
-	 */
-	public String getDescription()
+	public String getName() 
+	{
+		return name;
+	}
+	
+	public String getDescription() 
 	{
 		return description;
 	}
 	
 	/**
-	 * Creates this template in a target directory.
+	 * Adds a template to this descriptor.
+	 * It cannot be added more than once, and are exported/applied in the order added.
+	 * @param template the template to add.
+	 * @return this descriptor.
+	 */
+	public ProjectTemplate add(ProjectModule template)
+	{
+		if (!moduleSet.contains(template))
+			moduleList.add(template);
+		return this;
+	}
+
+	/**
+	 * Creates the templates in this descriptor in a target directory.
 	 * @param directory the directory.
 	 * @throws IOException if a problem happens while creating the template.
 	 */
 	public void createIn(File directory) throws IOException
 	{
-		if (directory.exists())
-		{
-			if (!directory.isDirectory())
-				throw new IOException("Target is not a directory: " + directory.getPath());
-			if (directory.listFiles().length == 0)
-				throw new IOException("Target is not an empty directory: " + directory.getPath());
-		}
-		
-		for (Entry e : entries)
-		{
-			String targetPath = directory.getPath() + File.separator + e.outputPath;
-			if (e.isDirectory())
-			{
-				if (!Common.createPath(targetPath))
-					throw new IOException("Could not create necessary directory: " + targetPath);
-			}
-			else
-			{
-				if (!Common.createPathForFile(targetPath))
-					throw new IOException("Could not create necessary directory for file: " + targetPath);
-				try (FileOutputStream fos = new FileOutputStream(targetPath, e.appending))
-				{
-					for (String resourcePath : e.resourcePaths)
-					{
-						try (InputStream in = Common.openResource(resourcePath))
-						{
-							if (in == null)
-								throw new IOException("INTERNAL ERROR: Could not find resource: " + resourcePath);
-							IOUtils.relay(in, fos);
-						}
-					}
-				}
-			}
-		}
+		for (ProjectModule template : moduleList)
+			template.createIn(directory);
+	}
+
+	@Override
+	public int compareTo(ProjectTemplate o)
+	{
+		return name.compareTo(o.name);
 	}
 	
 }